empty content utilization
