+++ src/components/molecules/EmptyListContent/EmptyListContent.styled.tsx
@@ -2,10 +2,12 @@ import {Space} from 'antd';
 
 import styled from 'styled-components';
 
-export const StyledEmptyTestSuitesDataContainer = styled(Space)`
+export const StyledEmptyListContainer = styled(Space)`
   display: flex;
   flex: 1;
   flex-direction: column;
   align-items: center;
   justify-content: center;
+
+  height: 100%;
 `;
