--- src/components/molecules/EmptyTestSuitesListContent/index.ts
@@ -1,1 +1,0 @@
-export {default} from './EmptyTestSuitesListContent';
