--- src/components/molecules/EmptyTestsListContent/EmptyTestsListContent.styled.tsx
@@ -1,11 +1,0 @@
-import {Space} from 'antd';
-
-import styled from 'styled-components';
-
-export const StyledEmptyTestsDataContainer = styled(Space)`
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-`;
