--- src/components/molecules/EmptyTestsListContent/index.ts
@@ -1,1 +1,0 @@
-export {default} from './EmptyTestsListContent';
