--- src/components/molecules/ExecutionStepIcon/ExecutionStepIcon.styled.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledExecutionStepIcon = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 20px;
-  height: 30px;
-`;
