--- src/components/molecules/ExecutionStepIcon/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ExecutionStepIcon';
