+++ src/components/molecules/index.ts
@@ -5,13 +5,10 @@ export {default as CLICommands} from './CLICommands';
 export {default as CopyCommand} from './CLICommands/CopyCommand';
 export {default as LabelsList} from './LabelsList';
 export {default as ExecutionStepsList} from './ExecutionStepsList';
-export {default as ExecutionStepIcon} from './ExecutionStepIcon';
 export {default as ExecutionName} from './ExecutionName';
 export {default as CookiesBanner} from './CookiesBanner';
 export {default as LogOutput} from './LogOutput';
 export {TestsVariablesList, ExecutionsVariablesList} from './Variables';
-export {default as EmptyTestsListContent} from './EmptyTestsListContent';
-export {default as EmptyTestSuitesListContent} from './EmptyTestSuitesListContent';
 export {default as EntityGrid} from './EntityGrid';
 export {default as notificationCall} from './Notification';
 export {default as EndpointModal} from './EndpointModal';
@@ -21,3 +18,4 @@ export {default as DragNDropList} from './DragNDropList';
 export {default as ArtifactsList} from './ArtifactsList';
 export {default as TestSuiteStepCard} from './TestSuiteStepCard';
 export {default as HelpCard} from './HelpCard';
+export {default as EmptyListContent} from './EmptyListContent';
