+++ src/components/organisms/EntityExecutions/EntityExecutionsContent/EmptyExecutionsListContent/index.ts
@@ -1,0 +1,1 @@
+export {default} from './EmptyExecutionsListContent';
