+++ src/components/organisms/EntityExecutions/EntityExecutionsContent/EntityExecutionsContent.styled.tsx
@@ -2,7 +2,17 @@ import {PageHeader, Tabs} from 'antd';
 
 import styled from 'styled-components';
 
-import Colors from '@styles/Colors';
+export const StyledContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 40px;
+  flex: 1;
+
+  height: 100%;
+  padding: 40px;
+
+  overflow-y: auto;
+`;
 
 export const StyledPageHeader = styled(PageHeader)`
   &.testkube-pageheader {
@@ -14,30 +24,6 @@ export const StyledPageHeader = styled(PageHeader)`
         margin: 0;
       }
     }
-  }
-`;
-
-export const SummaryGridWrapper = styled.div<{$gridCols: number}>`
-  display: grid;
-  grid-template-columns: repeat(${({$gridCols}) => $gridCols}, 1fr);
-  gap: 12px;
-`;
-
-export const SummaryGridItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-
-  height: 100px;
-  padding: 20px;
-  border-radius: 4px;
-
-  background: ${Colors.slate800};
-
-  transition: 0.3s;
-
-  &:hover {
-    background: ${Colors.slate800halfalpha};
   }
 `;
