+++ src/components/organisms/EntityExecutions/EntityExecutionsContent/SummaryGrid/index.ts
@@ -1,0 +1,1 @@
+export {default} from './SummaryGrid';
