+++ src/constants/entitiesConfig/EmptyEntitiesListContent/index.ts
@@ -1,0 +1,2 @@
+export {default as EmptyTestsListContent} from './EmptyTestsListContent';
+export {default as EmptyTestSuitesListContent} from './EmptyTestSuitesListContent';
