+++ src/constants/entitiesConfig/TestsConfig.tsx
@@ -12,11 +12,11 @@ import {
 
 import {Text} from '@custom-antd';
 
-import {EmptyTestsListContent} from '@molecules';
-
 import {useGetTestsQuery} from '@services/tests';
 
 import Colors from '@styles/Colors';
+
+import {EmptyTestsListContent} from './EmptyEntitiesListContent';
 
 const TestsPageDescription: React.FC = () => {
   return (
