--- src/models/creation.ts
@@ -1,5 +1,0 @@
-export type CreationBlueprintType = 'test';
-
-export type CreationBlueprintProps = {
-  entityType: CreationBlueprintType;
-};
