--- src/models/labels.ts
@@ -1,7 +1,0 @@
-import {EntityMap} from '@models/entityMap';
-
-interface LabelsState {
-  labelsObject: EntityMap;
-}
-
-export type {LabelsState};
