+++ src/models/test.ts
@@ -30,3 +30,24 @@ export type TestWithExecution = {
   test: Test;
   latestExecution?: Execution;
 };
+
+export type TestFilters = {
+  textSearch: string;
+  type: string;
+  pageSize: number;
+  page: number;
+  selector: string[];
+  createdAt: null;
+  status: Array<string>;
+};
+interface TestsState {
+  isLoading?: boolean;
+  dataList: Test[];
+  latestExecution: any;
+  filters: TestFilters;
+  totals: {};
+  filtered: {};
+  selectedTest: Test | null;
+}
+
+export type {TestsState};
