+++ src/models/testSuite.ts
@@ -48,3 +48,23 @@ export type TestSuiteWithExecution = {
   dataList: TestSuite;
   latestExecution?: TestSuiteExecution;
 };
+export type TestSuiteFilters = {
+  textSearch: string;
+  pageSize: number;
+  page: number;
+  selector: string[];
+  startDate: null;
+  endDate: null;
+  status: Array<string>;
+};
+interface TestSuitesState {
+  isLoading?: boolean;
+  dataList: TestSuite | {}[];
+  latestExecution?: Execution;
+  filters: TestSuiteFilters;
+  totals: {};
+  filtered: {};
+  selectedTestSuite?: TestSuite;
+}
+
+export type {TestSuitesState};
