+++ src/models/testSuiteExecution.ts
@@ -19,3 +19,13 @@ export type TestSuiteExecution = {
   labels?: EntityMap;
   stepResults?: TestSuiteStepExecutionResult[];
 };
+interface TestSuiteExecutionsState {
+  isLoading?: boolean;
+  dataList: TestSuiteExecution[];
+  filters: {textSearch?: string; pageSize?: number; page: 0; selector: string};
+  totals: {};
+  filtered: {};
+  selectedTestSuiteExecution?: TestSuiteExecution | null;
+}
+
+export type {TestSuiteExecutionsState};
